feat: improve http error management and rename jwt interceptor
