+++ front/src/app/core/core.providers.ts
@@ -1,9 +1,10 @@
 import { EnvironmentProviders, makeEnvironmentProviders } from '@angular/core';
 import { provideHttpClient, withInterceptors } from '@angular/common/http';
-import { jwtInterceptor } from './interceptors/jwt-interceptor.service';
+import { jwtInterceptor } from './interceptors/jwt-interceptor';
+import { httpErrorInterceptor } from './interceptors/http-error-interceptor';
 
 export function provideCore(): EnvironmentProviders {
   return makeEnvironmentProviders([
-    provideHttpClient(withInterceptors([jwtInterceptor]))
+    provideHttpClient(withInterceptors([jwtInterceptor,httpErrorInterceptor]))
   ]);
 }
