--- front/src/app/core/interceptors/jwt-interceptor.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { JwtInterceptorService } from './jwt-interceptor.service';
-
-describe('JwtInterceptorService', () => {
-  let service: JwtInterceptorService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(JwtInterceptorService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
